requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
typer==0.16.1
pydantic==2.11.7
pydantic-settings==2.10.1
python-dotenv==1.1.1
PyYAML==6.0.2
